Utilize AWS SDK waiters for instance state transitions
This change updates the AWS CPI to only use the built in waiting
functionality provided by the AWS SDK rather than the custom resource
waiting the we have built.

**Note**: One side effect is that we will likely no longer retry on the
`AbruptlyTerminated` error when creating instances. We should keep an
eye on this going forward to see if we should add more functionality.

[finishes #167798477](https://www.pivotaltracker.com/story/show/167798477)

Co-authored-by: Belinda Liu <bliu@pivotal.io>