Remove load balancer accessibility checks
This change removes the initial checks on the AWS API when load
balancers are specified in the vm cloud properties. We have little
confidence in these sanity checks and anticipate that further requests
will fail for a similar reason. By reducing the number of API requests
to AWS, we are hoping to reduce the chances of being rate limited.

Co-authored-by: James Myers <jmyers@pivotal.io>